requirements.txt
Flask==1.1.2
pyspark==3.1.2
azure-storage-blob==12.8.1
databricks-cli==0.14.3
requests==2.25.1
